+++ lib/pah/template.rb
@@ -57,6 +57,7 @@ apply_n :rspec
 apply_n :default
 apply_n :stylesheets
 apply_n :secure_headers
+apply_n :secret_token
 apply_n :omniauth
 apply_n :capybara
 apply_n :generators